requirements.txt
cloudpickle==1.6.0
future==0.18.2
gym==0.17.3
gym-retro==0.8.0
neat-python==0.92
numpy==1.19.2
opencv-python==4.4.0.44
pyglet==1.5.0
scipy==1.5.2
